fixes #3183 (Update hunspell to 1.7.1)

+++ libs/hunspell/src/affentry.hxx
@@ -1,7 +1,7 @@
 /* ***** BEGIN LICENSE BLOCK *****
  * Version: MPL 1.1/GPL 2.0/LGPL 2.1
  *
- * Copyright (C) 2002-2017 Németh László
+ * Copyright (C) 2002-2022 Németh László
  *
  * The contents of this file are subject to the Mozilla Public License Version
  * 1.1 (the "License"); you may not use this file except in compliance with
@@ -118,7 +118,7 @@ class PfxEntry : public AffEntry {
   const char* getKey() { return appnd.c_str(); }
   std::string add(const char* word, size_t len);
 
-  inline short getKeyLen() { return (short)appnd.size(); }
+  inline short getKeyLen() { return appnd.size(); }
 
   inline const char* getMorph() { return morphcode; }
 
@@ -199,7 +199,7 @@ class SfxEntry : public AffEntry {
   inline short getContLen() { return contclasslen; }
   inline const char* getAffix() { return appnd.c_str(); }
 
-  inline short getKeyLen() { return (short)appnd.size(); }
+  inline short getKeyLen() { return appnd.size(); }
 
   inline SfxEntry* getNext() { return next; }
   inline SfxEntry* getNextNE() { return nextne; }
